requirements.txt
requests
rich
tqdm
setuptools
python-dotenv
